+++ src/components/cell-control/autocompleter.jsx
@@ -2,21 +2,34 @@ import PropTypes from 'prop-types';
 import React, { forwardRef, useRef } from 'react';
 import CellControl from './cell-control.jsx';
 import AutocompleterControl from '../control/autocompleter.jsx';
+import styles from './select.css';
 
 const Autocompleter = forwardRef(function Autocompleter(props, ref) {
+  const classNames = {
+    container: undefined,
+    innerContainer: styles.container,
+    input: styles.input,
+    invalidInput: styles.invalidInput,
+    ...props.classNames,
+  };
   const refs = {
     container: useRef(),
   };
 
   return (
     <CellControl
-      className={props.className}
+      className={classNames.container}
       labelledBy={props.labelledBy}
       readOnly={props.readOnly}
       ref={refs.container}
     >
       <AutocompleterControl
         apiEndpoint={props.apiEndpoint}
+        classNames={{
+          container: classNames.innerContainer,
+          input: styles.input,
+          invalidInput: styles.invalidInput,
+        }}
         displayValueEvaluator={props.displayValueEvaluator}
         id={props.id}
         labelledBy={props.labelledBy}
@@ -29,6 +42,7 @@ const Autocompleter = forwardRef(function Autocompleter(props, ref) {
         required={props.required}
         searchParam={props.searchParam}
         validationMessage={props.validationMessage}
+        validationMessageTooltip
         value={props.value}
         wrapperRef={refs.container}
       />
@@ -39,8 +53,12 @@ const Autocompleter = forwardRef(function Autocompleter(props, ref) {
 Autocompleter.propTypes = {
   /** `apiEndpoint` should be the route of the api's endpoint (excluding the base api), eg. `/workspaces`. */
   apiEndpoint: PropTypes.string,
-  /** A class name for the table cell container. */
-  className: PropTypes.string,
+  classNames: PropTypes.shape({
+    /** A class name for the table cell container. */
+    container: PropTypes.string,
+    /** A class name for the nested select container. */
+    innerContainer: PropTypes.string,
+  }),
   /** displayValueEvaluator is handled if the key following: `title`, `name`, `full_name`, `currency`; Otherwise, pass in something like `displayValueEvaluator: (model) -> { model.rate_card_name }` */
   displayValueEvaluator: PropTypes.func,
   id: PropTypes.string.isRequired,
@@ -62,7 +80,7 @@ Autocompleter.propTypes = {
 
 Autocompleter.defaultProps = {
   apiEndpoint: undefined,
-  className: undefined,
+  classNames: {},
   displayValueEvaluator: undefined,
   models: [],
   onChange: () => {},