+++ styleguide.config.js
@@ -43,7 +43,9 @@ module.exports = {
     components: [
       path.join(__dirname, 'src/components/table/table.jsx'),
       path.join(__dirname, 'src/components/cell-control/autocompleter.jsx'),
+      path.join(__dirname, 'src/components/cell-control/choice.jsx'),
       path.join(__dirname, 'src/components/cell-control/input.jsx'),
+      path.join(__dirname, 'src/components/cell-control/number.jsx'),
       path.join(__dirname, 'src/components/cell-control/select.jsx'),
       path.join(__dirname, 'src/components/cell-control/textarea.jsx'),
     ],