+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - chore(control): deprecate Control component, move validationMessage to FormControl
 - chore(all): reference invalid icon as description for interactive elements
 - feat(autocomplete cell): update styling to match new cell designs
+- feat(choice cell): implement custom field single choice field cell component (words)
 
 ## 0.96.1